picme/requirements.txt
appdirs==1.4.3
Django==1.10.6
django-braces==1.10.0
django-guardian==1.4.6
django-model-utils==2.6.1
django-userena==2.0.1
easy-thumbnails==2.3
geoip2==2.4.2
html2text==2014.12.29
maxminddb==1.3.0
olefile==0.44
packaging==16.8
Pillow==4.0.0
pyparsing==2.2.0
requests==2.13.0
six==1.10.0
websockets==3.2
